requirements.txt
arrow==1.3.0
cachetools==5.3.2
cbor2==5.6.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
easyocr==1.7.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-vision==3.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
luma.core==2.4.2
luma.oled==3.13.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycrypto==2.6.1
pyftdi==0.55.0
pyparsing==3.1.1
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
requests==2.31.0
RPi.GPIO==0.7.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.1
shapely==2.0.2
six==1.16.0
smbus2==0.4.3
spidev==3.6
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.0.238
urllib3==2.1.0
